+++ test/unit/lib/fs/package-json/read_spec.js
@@ -19,7 +19,7 @@ describe('lib/fs/package-json: read', function () {
             fs.mkdirSync(join(packagePath.name, '.git'));
             fs.writeFileSync(join(packagePath.name, '.DS_Store'), '');
 
-            packageJSON.read.all(packagePath.name)
+            packageJSON.readPackages(packagePath.name)
                 .then(function (pkgs) {
                     pkgs.should.eql({
                         casper: {
@@ -50,7 +50,7 @@ describe('lib/fs/package-json: read', function () {
             fs.writeFileSync(join(packagePath.name, 'testtheme', 'package.json'), pkgJson);
             fs.writeFileSync(join(packagePath.name, 'testtheme', 'index.hbs'), '');
 
-            packageJSON.read.all(packagePath.name)
+            packageJSON.readPackages(packagePath.name)
                 .then(function (pkgs) {
                     pkgs.should.eql({
                         testtheme: {
@@ -83,7 +83,7 @@ describe('lib/fs/package-json: read', function () {
             fs.writeFileSync(join(packagePath.name, 'testtheme', 'package.json'), pkgJson);
             fs.writeFileSync(join(packagePath.name, 'testtheme', 'index.hbs'), '');
 
-            packageJSON.read.all(packagePath.name)
+            packageJSON.readPackages(packagePath.name)
                 .then(function (pkgs) {
                     pkgs.should.eql({
                         testtheme: {
@@ -114,7 +114,7 @@ describe('lib/fs/package-json: read', function () {
             fs.mkdirSync(join(packagePath.name, '.git'));
             fs.writeFileSync(join(packagePath.name, '.DS_Store'), '');
 
-            packageJSON.read.one(packagePath.name, 'casper')
+            packageJSON.readPackage(packagePath.name, 'casper')
                 .then(function (pkgs) {
                     pkgs.should.eql({
                         casper: {
@@ -145,7 +145,7 @@ describe('lib/fs/package-json: read', function () {
             fs.writeFileSync(join(packagePath.name, 'testtheme', 'package.json'), pkgJson);
             fs.writeFileSync(join(packagePath.name, 'testtheme', 'index.hbs'), '');
 
-            packageJSON.read.one(packagePath.name, 'testtheme')
+            packageJSON.readPackage(packagePath.name, 'testtheme')
                 .then(function (pkgs) {
                     pkgs.should.eql({
                         testtheme: {
@@ -178,7 +178,7 @@ describe('lib/fs/package-json: read', function () {
             fs.writeFileSync(join(packagePath.name, 'testtheme', 'package.json'), pkgJson);
             fs.writeFileSync(join(packagePath.name, 'testtheme', 'index.hbs'), '');
 
-            packageJSON.read.one(packagePath.name, 'testtheme')
+            packageJSON.readPackage(packagePath.name, 'testtheme')
                 .then(function (pkgs) {
                     pkgs.should.eql({
                         testtheme: {
@@ -209,7 +209,7 @@ describe('lib/fs/package-json: read', function () {
             fs.mkdirSync(join(packagePath.name, 'not-casper'));
             fs.writeFileSync(join(packagePath.name, 'not-casper', 'index.hbs'), '');
 
-            packageJSON.read.one(packagePath.name, 'casper')
+            packageJSON.readPackage(packagePath.name, 'casper')
                 .then(function (pkgs) {
                     pkgs.should.eql({
                         casper: {
@@ -232,7 +232,7 @@ describe('lib/fs/package-json: read', function () {
             fs.writeFileSync(join(packagePath.name, 'casper.zip'), '');
             fs.writeFileSync(join(packagePath.name, '.DS_Store'), '');
 
-            packageJSON.read.one(packagePath.name, 'casper')
+            packageJSON.readPackage(packagePath.name, 'casper')
                 .then(function () {
                     done('Should have thrown an error');
                 })
@@ -250,7 +250,7 @@ describe('lib/fs/package-json: read', function () {
             fs.writeFileSync(join(packagePath.name, 'casper.zip'), '');
             fs.writeFileSync(join(packagePath.name, '.DS_Store'), '');
 
-            packageJSON.read.one(packagePath.name, 'casper.zip')
+            packageJSON.readPackage(packagePath.name, 'casper.zip')
                 .then(function (pkg) {
                     pkg.should.eql({});
